Refactor & write some jsDocs
